% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleStats.R
\name{linear_regression_test}
\alias{linear_regression_test}
\title{Parameter Uji Linier Regresi}
\usage{
linear_regression_test(x, y)
}
\arguments{
\item{x}{Vektor numerik yang berisi nilai variabel independen (prediktor)}

\item{y}{Vektor numerik yang berisi nilai variabel dependen (respon)}
}
\value{
Intercept, Slope, R Squared, P-Value
example
x <- c(1, 2, 3, 4, 5)
y <- c(2, 4, 5, 4, 5)
}
\description{
Parameter Uji Linier Regresi
}
