% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleStats.R
\name{logistic_regression_test}
\alias{logistic_regression_test}
\title{Parameter Uji regresi logistik}
\usage{
logistic_regression_test(x, y)
}
\arguments{
\item{x}{Vektor numerik yang berisi nilai variabel independen (prediktor)}

\item{y}{Vektor biner (0/1) yang berisi nilai variabel dependen (respon)}
}
\value{
Intercept, Slope, R Squared, P-Value
example
x <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
y <- c(0, 0, 0, 0, 1, 1, 1, 1, 1, 1)
}
\description{
Parameter Uji regresi logistik
}
