% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleStats.R
\name{one_sample_t_test}
\alias{one_sample_t_test}
\title{Parameter One-Sample T-Test}
\usage{
one_sample_t_test(data, mu_0)
}
\arguments{
\item{data}{data yang digunakan untuk uji one sample t test}

\item{mu_0}{rata-rata populasi yang dijadikan acuan dalam pengujian hipotesis}
}
\value{
nilai t hitung
}
\description{
Parameter One-Sample T-Test
}
\examples{
data <- c(140,155,160,145,150,138,152,148,157,143)
mu_0 <- 150
}
