% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleStats.R
\name{two_sample_t_test}
\alias{two_sample_t_test}
\title{Parameter Two-Sample T-Test}
\usage{
two_sample_t_test(data1, data2)
}
\arguments{
\item{data1}{data pertama yang digunakan untuk uji Two sample t test}

\item{data2}{data kedua yang digunakan untuk uji Two sample t test}
}
\value{
nilai t hitung
}
\description{
Parameter Two-Sample T-Test
}
\examples{
data1 <- c(78,85,88,92,76,81,90,87,84,79)
data2 <- c(72,80,75,78,74,79,77,81,76,73)
}
